Share the value-class-null-enum test in PostgresNullBindingContract
The jdbc and r2dbc versions asserted identical behavior with only
differing explanatory comments about internal bindNull mechanics, so
they belonged in the shared contract like the rest of the pair. Only
the typed-null escape hatch (module-specific API) stays on the
concrete subclasses.
